requirements.txt
Django==2.1.2
django-cors-headers==2.4.0
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2-binary
dj_database_url
